OCR/requirements.txt
fastapi
uvicorn
pytesseract
pillow
python-multipart
opencv-python-headless
numpy
pdf2image
pillow-heif